requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.2.3
cycler==0.10.0
dataclasses==0.8
Flask==1.1.2
Flask-BasicAuth==0.2.0
flatbuffers==1.12
gast==0.3.3
gevent==20.12.1
geventhttpclient==1.4.4
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
locust==1.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.0
onnx==1.8.0
onnxruntime==1.2.0
opencv-python==4.1.1.26
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.1.0
pip==20.3.3
pkg-resources==0.0.0
protobuf==3.14.0
psutil==5.8.0
ptflops==0.6.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
setuptools==51.1.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.1
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.8.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
